% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illumina_manifests.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear Cached Manifests}
\usage{
clear_cache(chip = NULL, verbose = TRUE, ask = TRUE)
}
\arguments{
\item{chip}{Character string specifying which chip's cache to clear,
or \code{NULL} to clear all cached manifests. Default is \code{NULL}.}

\item{verbose}{Logical. Print messages. Default is \code{TRUE}.}

\item{ask}{Ask for permission to download or delete the cache. Default is \code{TRUE}.}
}
\value{
Invisibly returns a character vector of deleted paths.
}
\description{
Removes cleaned manifest files from the slideimp data directory.
}
\examples{
\dontrun{
clear_cache("EPICv2")
clear_cache()
}
}
