% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_matrix.R
\name{dedup_matrix}
\alias{dedup_matrix}
\title{De-duplicate Matrix by CpG Names}
\usage{
dedup_matrix(obj, chip, method = c("mean", "median"), verbose = TRUE)
}
\arguments{
\item{obj}{A numeric matrix with unique column and row names.}

\item{chip}{Either "MSA", "EPICv2", or a data.frame with \code{IlmnID} and \code{Name} columns.}

\item{method}{Aggregation method for duplicates: "mean" or "median".}

\item{verbose}{Logical; if TRUE, print informative messages.}
}
\value{
A de-duplicated matrix.
}
\description{
The \code{EPICv2} and \code{MSA} chips can return duplicated CpG names. This function de-duplicate
the beta matrix.
}
