% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illumina_manifests.R
\name{get_manifest}
\alias{get_manifest}
\title{Retrieve Cleaned Illumina Methylation Manifest}
\usage{
get_manifest(
  chip = NULL,
  rawdir = withr::local_tempdir(),
  force = FALSE,
  clean_up = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{chip}{The name of the Illumina methylation chip. If \code{NULL}, then all available
options are returned}

\item{rawdir}{Directory where raw manifest files are downloaded and stored.
Defaults to a temporary directory.}

\item{force}{Forces re-download and re-cleaning of the manifest.
Default to \code{FALSE}.}

\item{clean_up}{deletes the raw manifest file after cleaning. Useful if
\code{rawdir} is not a temporary folder. Default to \code{FALSE}.}

\item{verbose}{prints messages. Default is \code{TRUE}.}
}
\value{
The path to the cleaned manifest file as a character string,
or \code{invisible(NULL)} if the chip is invalid.
}
\description{
Retrieves the cleaned manifest for the specified Illumina methylation array chip.
If not already cleaned, it downloads the raw manifest, applies the appropriate
cleaning function, and stores the result as an .fst file.
}
\examples{
\dontrun{
get_manifest("450K")
}
}
