% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideimp_path.R
\name{load_data_function}
\alias{load_data_function}
\title{Dependency Factory}
\usage{
load_data_function(object_name, object_hash)
}
\arguments{
\item{object_name}{name of object}

\item{object_hash}{hash of object}
}
\description{
This function factory generate functions to load in these data based on a
template.
}
\keyword{internal}
