% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideimp_path.R
\name{set_slideimp_path}
\alias{set_slideimp_path}
\title{Set Path for slideimp Data}
\usage{
set_slideimp_path(path)
}
\arguments{
\item{path}{Character string specifying the directory path, or \code{NULL} to
reset to default.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Sets the slideimp data directory path for the current R session.
}
\examples{
# default path
get_slideimp_path()

# set path for this session
set_slideimp_path("test")
get_slideimp_path()

# reset to default
set_slideimp_path(NULL)
get_slideimp_path()
}
