% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideimp_resolve_group.character.R
\name{slideimp_resolve_group.character}
\alias{slideimp_resolve_group.character}
\title{Resolve a chip name to a manifest data.frame}
\usage{
\method{slideimp_resolve_group}{character}(x)
}
\arguments{
\item{x}{A character scalar naming a supported chip (see
\code{slideimp_arrays}).}
}
\value{
A data.frame with \code{feature} and related manifest columns.
}
\description{
Method for \code{\link[slideimp:slideimp_resolve_group]{slideimp::slideimp_resolve_group()}} that handles character
input by looking up the corresponding Illumina manifest.
}
\examples{

\dontrun{
slideimp::slideimp_resolve_group("EPICv2")
}

}
\keyword{internal}
