% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illumina_manifests.R
\name{download_manifest}
\alias{download_manifest}
\title{Download Raw Illumina Methylation Manifest}
\usage{
download_manifest(chip = NULL, rawdir = withr::local_tempdir(), verbose = TRUE)
}
\arguments{
\item{chip}{The name of the Illumina methylation chip. If \code{NULL}, then all available
options are returned}

\item{rawdir}{Directory where raw manifest files are downloaded and stored.
Defaults to a temporary directory.}

\item{verbose}{prints messages. Default is \code{TRUE}.}
}
\value{
The path to the downloaded manifest file as a character string,
or \code{invisible(NULL)} if the chip is invalid or not specified.
}
\description{
Downloads the raw manifest file for a specified Illumina methylation array chip.
It checks for existing files and verifies integrity using a pre-defined hash.
If the file is missing or the hash mismatches, it re-downloads the file.
}
\examples{
\dontrun{
download_manifest("EPICv1")
}
}
