% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illumina_manifests.R
\name{ilmn_manifest}
\alias{ilmn_manifest}
\title{Load Illumina Methylation Manifest Data}
\usage{
ilmn_manifest(
  chip = NULL,
  dedupped = FALSE,
  rawdir = withr::local_tempdir(),
  force = FALSE,
  clean_up = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{chip}{The name of the Illumina methylation chip. If \code{NULL}, then all available
options are returned}

\item{dedupped}{Use deduplicated probe names for EPICv2 and MSA chips (\code{TRUE})
or IlmnID (\code{FALSE}). Default is \code{FALSE}.}

\item{rawdir}{Directory where raw manifest files are downloaded and stored.
Defaults to a temporary directory.}

\item{force}{Forces re-download and re-cleaning of the manifest.
Default to \code{FALSE}.}

\item{clean_up}{deletes the raw manifest file after cleaning. Useful if
\code{rawdir} is not a temporary folder. Default to \code{FALSE}.}

\item{verbose}{prints messages. Default is \code{TRUE}.}

\item{...}{Additional arguments passed to \code{\link[fst:write_fst]{fst::read_fst()}} for reading the
cleaned file.}
}
\value{
A \code{data.frame()} with columns "feature_id" (probe identifiers) and
"group" (chromosomal locations), or \code{invisible(NULL)} if the chip is invalid.
}
\description{
Loads the cleaned manifest data for a specified Illumina methylation array chip,
returning a unique data.frame with feature identifiers and their corresponding
chromosomal groups.
}
\examples{

ilmn_manifest()

\dontrun{
ilmn_manifest("EPICv2", dedupped = TRUE)
}
}
