% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideimp.extra-package.R
\docType{package}
\name{slideimp.extra-package}
\alias{slideimp.extra}
\alias{slideimp.extra-package}
\title{slideimp.extra: slideimp Package Helpers}
\description{
Contains helper functions for the slideimp package. These functions handle the pathing, download the Illumina Methylation manifests, clean them, and store the cleaned objects locally.
}
\author{
\strong{Maintainer}: Hung Pham \email{amser.hoanghung@gmail.com}

}
\keyword{internal}
